Removed client side school renaming since it's done server side now
